move location normalisation in router.sync.normalise
